rename folder shared for user
