+++ resources/js/shared/auth/forogot-password.tsx
