+++ resources/js/shared/auth/register.tsx
